renamed logger/types.ts to logger/logger-types.ts
