+++ flow-core/src/utils/logger/index.ts
@@ -10,9 +10,9 @@ export type {
   LoggerConfig,
   ErrorSeverity,
   ErrorHandlingOptions,
-} from './types.ts';
+} from './logger-types.ts';
 
-export { LogLevel } from './types.ts';
+export { LogLevel } from './logger-types.ts';
 
 // Export formatting utilities
 export type { LogLevel as LogLevelString } from './formatters.ts';
@@ -60,7 +60,7 @@ export {
 } from './error-handlers.ts';
 
 // Import types for function signatures
-import type { StructuredLogger } from './types.ts';
+import type { StructuredLogger } from './logger-types.ts';
 import type { EnhancedStructuredLogger } from './structured-logger.ts';
 import { createLogger, createEnhancedLogger } from './structured-logger.ts';
