+++ flow-core/tests/integration/logger/formatters.test.ts
@@ -10,7 +10,7 @@ import {
   createContextSummary,
   mergeLogContext
 } from '../../../src/utils/logger/index.ts';
-import type { LogContext } from '../../../src/utils/logger/types.ts';
+import type { LogContext } from '../../../src/utils/logger/logger-types.ts';
 
 console.log('🧪 Testing flow-core formatting functions...\n');
